components #42 updates 12/15
