+++ client/public/elements/pages/components/app-components.tpl.js
@@ -19,7 +19,7 @@ import "../../components/pagination"
 import "../../components/person-preview"
 import "../../components/quick-search"
 import "../../components/search"
-import "../../components/text-field"
+import "../../components/edit-profile"
 import "../../components/view-all"
 import "../../components/work-preview"
 
@@ -101,58 +101,7 @@ return html`
     align-items: center;
     height: 100px;
   }
-  .wrapper {
-    display: grid;
-    grid-template-columns: 45% 45% 5% 5%;
-    grid-gap: .2%;
-    background-color: #fff;
-    color: #444;
-  }
   
-#rearrange{
-    text-align: center;
-}
-  
-#delete{
-    vertical-align: center
-}
-  
-label{
-    text-align: left;
-    vertical-align: middle;
-}
-
-  
-#c2{
-    width: 100%;
-    margin: 3% 0 3%;
-}
-  
-#textbox {
-    padding: 2% 0 2%;
-    display: flex;
-    width: 100%;
-    box-sizing: border-box;
-}
-  
-#radiobox {
-    padding: 2% 0 2%;
-    box-sizing: border-box;
-    display: flex;
-}
-  
-#flexible-width {
-    flex: 1; /* my goal is that the width always fills up independent of browser width */
-    border: 5px;
-}
-#flexible-width-radio { 
-    border: 5px;
-    flex: 1;    
-}     
-input[type=text]{
-    box-shadow: inset 0 0 1.5px black, 0 0 2px grey;
-    padding: 1.5%;
-}      
   ${styles}
 </style>
 
@@ -446,7 +395,7 @@ There are three slots to populate the hero content - "top", "main", and "bottom"
 </section>
 
 <section>
-<h2>Profile About Editor</h2>
+<h1>Profile About Editor</h1>
 <rp-profile-about-editor></rp-profile-about-editor>
 </section>
